+++ lib/fields/floor.dart
@@ -8,6 +8,7 @@ import 'package:flutter/material.dart';
 import 'package:every_door/models/field.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:logging/logging.dart';
 
 import '../widgets/radio_field.dart';
 
@@ -64,7 +65,7 @@ class _FloorInputFieldState extends ConsumerState<FloorInputField> {
       floors = await osmData.getFloorsAround(widget.element.location, addr);
       Floor.collapseList(floors);
     } on Exception catch (e) {
-      print(e);
+      Logger('FloorInputField').warning('Error getting floors', e);
       floors = [];
     }
     final currentFloor = Floor.fromTags(tags);