+++ lib/providers/legend.dart
@@ -2,7 +2,6 @@ import 'package:every_door/helpers/good_tags.dart';
 import 'package:every_door/models/amenity.dart';
 import 'package:every_door/models/preset.dart';
 import 'package:every_door/providers/presets.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';