+++ lib/screens/settings.dart
@@ -9,6 +9,7 @@ import 'package:every_door/providers/presets.dart';
 import 'package:every_door/screens/settings/account.dart';
 import 'package:every_door/screens/settings/changes.dart';
 import 'package:every_door/screens/settings/imagery.dart';
+import 'package:every_door/screens/settings/log.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_dropdown_alert/alert_controller.dart';
 import 'package:flutter_dropdown_alert/model/data_alert.dart';
@@ -170,9 +171,17 @@ class VersionSection extends AbstractSettingsSection {
     return Padding(
       padding: const EdgeInsets.only(top: 10.0),
       child: Center(
-        child: Text(
-          'Every Door $kAppVersion',
-          style: TextStyle(color: Colors.grey, fontSize: 12.0),
+        child: GestureDetector(
+          child: Text(
+            'Every Door $kAppVersion',
+            style: TextStyle(color: Colors.grey, fontSize: 12.0),
+          ),
+          onTap: () {
+            Navigator.push(
+              context,
+              MaterialPageRoute(builder: (context) => LogDisplayPage()),
+            );
+          },
         ),
       ),
     );